Move egraph implementation into egraphs/basic/ subfolder
egraphs/ now has a clean separation:
- egraphs/traits.rs: public interface (EgraphTrait, Conflict, EgraphResult)
- egraphs/basic/: implementation (Egraph struct, CC, proof forest, repr)
- egraphs/mod.rs: re-exports for backward-compatible paths
