+++ src/egraphs/basic/congruence_closure.rs
