+++ src/egraphs/basic/datastructures.rs
