+++ src/egraphs/basic/egraph.rs
