+++ src/egraphs/basic/mod.rs
@@ -1,0 +1,10 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+pub(crate) mod congruence_closure;
+pub(crate) mod datastructures;
+pub mod egraph;
+pub(crate) mod proofforest;
+pub mod repr;
+pub(crate) mod unionfind;
+pub(crate) mod utils;
