+++ src/egraphs/basic/proofforest.rs
