+++ src/egraphs/basic/repr.rs
