+++ src/egraphs/basic/utils.rs
